#164103119 Bugs and errors

+++ app/my_api/v2/party.py
@@ -35,9 +35,9 @@ class CreatePartyV2(Resource):
     @jwt_required
     def post(self):
         data = CreatePartyV2.parser.parse_args()
-        name = ['name']
-        hqaddress = ['hqaddress']
-        logourl = ['logourl']
+        name = data['name']
+        hqaddress = data['hqaddress']
+        logourl = data['logourl']
         # validation
         while True:
             if not name or not hqaddress or not logourl:
@@ -66,15 +66,27 @@ class CreatePartyV2(Resource):
                 })
                 party_exist = cur.fetchone()
                 if party_exist is None:
-                    cur.execute("INSERT INTO Parties(name, hqaddress, logourl) VALUES(%(name)s, %(hqaddress)s, %(logourl)s);", {
+                    cur.execute("INSERT INTO Parties(name, hqaddress, logourl) VALUES(%(name)s, %(hqaddress)s, %(logourl)s) RETURNING party_id;", {
                         'name': data['name'], 'hqaddress': data['hqaddress'], 'logourl': data['logourl']
                     })
                     connection.commit()
-                    return {'Message': 'Party successfully added'}, 200
+                    new_id = cur.fetchone()
+                    item = {
+                        'id': new_id,
+                        'name': name
+                    }
+                    return {
+                        'status': 201,
+                        'Message': 'Party successfully added',
+                        'data': item}, 201
                 else:
-                    return {'Message': 'Party already exists'}, 409
+                    return {
+                        'status': 409,
+                        'Message': 'Party already exists'}, 409
             else:
-                return {'Message': 'This panel is for administrators only'}, 403
+                return {
+                    'status': 403,
+                    'Message': 'This panel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -88,7 +100,10 @@ class GetPartiesV2(Resource):
         try:
             cur.execute("SELECT * FROM Parties;")
             parties = cur.fetchall()
-            return {'Message': parties}, 200
+            return {
+                'status': 200,
+                'Message': 'This includes all the parties',
+                'data': parties}, 200
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -116,7 +131,10 @@ class EditPartyV2(Resource):
         try:
             cur.execute("SELECT * FROM Parties WHERE party_id = %s", [party_id])
             party = cur.fetchall()
-            return {'Message': party}, 200
+            return {
+                'status': 200,
+                'Message': 'This are the details of your requested party',
+                'data': party}, 200
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -145,9 +163,14 @@ class EditPartyV2(Resource):
             if is_admin:
                 cur.execute("UPDATE Parties SET name = %s, hqaddress = %s WHERE party_id = %s", (name, hqaddress, party_id))
                 connection.commit()
-                return {'Message': 'The party was updated'}, 200
+                return {
+                    'status': 200,
+                    'Message': 'The party was updated',
+                    'data': data}, 200
             else:
-                return {'Message': 'This pannel is for administrators only'}, 403
+                return {
+                    'status': 403,
+                    'Message': 'This pannel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -163,10 +186,21 @@ class EditPartyV2(Resource):
             user_exists = cur.fetchone()
             is_admin = user_exists[0]
             if is_admin:
-                cur.execute("DELETE FROM Parties WHERE party_id = %s;", [party_id])
-                return {'Message': 'Party deleted'}, 200
+                cur.execute("SELECT * FROM Parties WHERE party_id = %s;", [party_id])
+                exists_party = cur.fetchone()
+                if not exists_party:
+                    return {
+                        'status': 404,
+                        'Message': 'We cannot find office by that id'}, 404
+                else:
+                    cur.execute("DELETE FROM Parties WHERE party_id = %s;", [party_id])
+                    return {
+                        'status': 200,
+                        'Message': 'Party deleted'}, 200
             else:
-                return {'Message': 'This pannel is for administrators only'}, 403
+                return {
+                    'status': 403,
+                    'Message': 'This pannel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
